+++ hydra-cluster/src/Hydra/Cluster/Fixture.hs
@@ -63,4 +63,5 @@ data KnownNetwork
   = Preview
   | Preproduction
   | Mainnet
-  deriving stock (Show)
+  deriving stock (Generic, Show, Eq, Enum, Bounded)
+  deriving anyclass (ToJSON, FromJSON)